Updates from Code Review
