+++ src/components/Content/index.tsx
@@ -3,97 +3,8 @@ import { htmlToSteamBBCode, cleanBBCode } from "util/htmlToSteamBBCode";
 import { trackError, trackEvent } from "util/analytics";
 import ReviewEditor from "components/Editor";
 import Preview from "components/Preview";
+import { normalizePreviewSpoilers } from "./normalizePreviewSpoilers";
 import "./content.scss";
-
-function isSpoilerElement(node: Node | null): node is HTMLElement {
-  return node instanceof HTMLElement && node.tagName === "SPAN" && node.getAttribute("data-type") === "spoiler";
-}
-
-function normalizePreviewSpoilers(html: string): string {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, "text/html");
-
-  const inlineSelectors = ["strong", "b", "em", "i", "u", "s", "strike"];
-
-  doc.querySelectorAll(inlineSelectors.map(tag => `${tag} > span[data-type='spoiler']`).join(", ")).forEach(node => {
-    const spoiler = node as HTMLElement;
-    const inlineWrapper = spoiler.parentElement;
-
-    if (!inlineWrapper || inlineWrapper.childNodes.length !== 1) {
-      return;
-    }
-
-    const mergedSpoiler = spoiler.cloneNode(false) as HTMLElement;
-    inlineWrapper.replaceWith(mergedSpoiler);
-    mergedSpoiler.appendChild(inlineWrapper);
-  });
-
-  doc.querySelectorAll("span[data-type='spoiler']").forEach(node => {
-    const spoiler = node as HTMLElement;
-    const parentSpoiler = spoiler.parentElement?.closest("span[data-type='spoiler']");
-
-    if (!parentSpoiler) {
-      return;
-    }
-
-    while (spoiler.firstChild) {
-      spoiler.parentNode?.insertBefore(spoiler.firstChild, spoiler);
-    }
-
-    spoiler.remove();
-  });
-
-  const mergeAdjacentSpoilers = (parent: ParentNode) => {
-    let current = parent.firstChild;
-
-    while (current) {
-      if (current instanceof HTMLElement) {
-        mergeAdjacentSpoilers(current);
-      }
-
-      if (!isSpoilerElement(current)) {
-        current = current?.nextSibling ?? null;
-        continue;
-      }
-
-      let next = current.nextSibling;
-
-      while (
-        next?.nodeType === Node.TEXT_NODE &&
-        !(next.textContent ?? "").trim() &&
-        isSpoilerElement(next.nextSibling)
-      ) {
-        current.appendChild(next);
-        next = current.nextSibling;
-      }
-
-      while (isSpoilerElement(next)) {
-        while (next.firstChild) {
-          current.appendChild(next.firstChild);
-        }
-
-        const nodeToRemove = next;
-        next = next.nextSibling;
-        nodeToRemove.remove();
-
-        while (
-          next?.nodeType === Node.TEXT_NODE &&
-          !(next.textContent ?? "").trim() &&
-          isSpoilerElement(next.nextSibling)
-        ) {
-          current.appendChild(next);
-          next = current.nextSibling;
-        }
-      }
-
-      current = next;
-    }
-  };
-
-  mergeAdjacentSpoilers(doc.body);
-
-  return doc.body.innerHTML;
-}
 
 export type AppContextType = null | {
   setHTMLContent: React.Dispatch<React.SetStateAction<string>>;
